chore(build): clear the warnings the Kotlin 2.3.20 upgrade exposed
Follow-up sweep on the 2.3.20 bump. No behaviour changes.

Dependency alignment:
- coil3 3.0.4 -> 3.4.0. Not 3.5.0: its iOS klib is abi_version=2.4.0 (Kotlin
  2.4.0 compiler), which 2.3.20's resolver skips with exactly the same misleading
  "Could not find <klib>" error that started this whole upgrade. 3.4.0 is
  abi_version=2.3.0 and is the newest release Kotlin 2.3.20 can consume.
- navigation-compose-kmp 2.9.0-beta02 -> 2.9.2 (stable release of the same line;
  the beta was never revisited after Phase 5).

Warning fixes:
- Scoped the `kotlinx.cinterop.ExperimentalForeignApi` opt-in to the native source
  sets in :shared and :data-firebase. It was applied via `sourceSets.all`, so it
  also hit Android/JVM source sets where the annotation class is not on the
  classpath, warning "Opt-in requirement marker ... is unresolved" on every
  compile*KotlinAndroid task.
- Removed 9 redundant null-handling constructs in shared/commonMain that Kotlin
  2.3.20's stricter analysis flags: 2 unnecessary safe calls, 2 dead elvis
  branches, 4 unnecessary !! on already-smart-cast values, 1 redundant .toInt().
  Two of the dead elvis branches were also hardcoded UI strings ("SSB Aspirant",
  "") that could never be reached.

Deliberately NOT done, with reasons:
- shared/.../ui/premium/UpgradeScreen.kt:252 still warns about a redundant `else`.
  Left alone: a parallel payment-hardening session owns that area.
- Compose Multiplatform 1.11 deprecates the `compose.runtime` / `compose.material3`
  / `compose.uiTest` DSL accessors in favour of direct GAVs. Not migrated: the
  accessors resolve material3 to its own 1.9.0 line and material-icons-extended to
  1.7.3, i.e. the plugin maintains a curated version mapping that hand-written
  catalog entries would freeze. Migrating now would silently pin those on the next
  CMP bump. Revisit when CMP publishes a matching version-catalog bundle.
- The "Skiko dependencies' versions are incompatible" warning is unavoidable until
  Kotlin 2.4.x lets coil3 move to 3.5.0. Documented in the catalog.

Verified: ./gradlew check green (including :shared:iosSimulatorArm64Test),
:shared:compileKotlinIosArm64, :data-firebase:linkReleaseFrameworkIosArm64,
validate-security.sh 10/10.

+++ shared/src/commonMain/kotlin/com/ssbmax/shared/ui/interviewsession/InterviewSessionScreen.kt
@@ -221,7 +221,7 @@ private fun ErrorContent(error: String?, onRetry: () -> Unit) = Column(
 private fun InterviewContent(uiState: InterviewSessionUiState, viewModel: InterviewSessionViewModel) {
     val progressDescription = stringResource(
         Res.string.interview_progress_content_description,
-        uiState.getProgressPercentage().toInt().coerceIn(0, 100)
+        uiState.getProgressPercentage().coerceIn(0, 100)
     )
     Column(
         modifier = Modifier.fillMaxSize().padding(16.dp).verticalScroll(rememberScrollState()),
